third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCPeerConnection-canTrickleIceCandidates-expected.txt

This is a testharness.js-based test.
FAIL canTrickleIceCandidates property is null prior to setRemoteDescriptionassert_equals: canTrickleIceCandidates property is null expected (object) null but got (undefined) undefined
FAIL canTrickleIceCandidates property is true after setRemoteDescription with a=ice-options:tricklepromise_test: Unhandled rejection with value: object "OperationError: TEST_ERROR"
FAIL canTrickleIceCandidates property is false after setRemoteDescription without a=ice-options:tricklepromise_test: Unhandled rejection with value: object "OperationError: TEST_ERROR"
Harness: the test ran to completion.
